+++ amdx8664/xopOpcodeMap10.go
@@ -1,6 +1,6 @@
 package amdx8664
 
-func xopOpcodeMap10(curByte int) Instruction {
+func xopOpcodeMap10(curByte byte) Instruction {
 	switch curByte {
 	case 0x0:
 		return ADD
